+++ duckscript_sdk/src/sdk/std/collections/mod.rs
@@ -10,6 +10,7 @@ mod is_map;
 mod map;
 mod map_get;
 mod map_is_empty;
+mod map_load_properties;
 mod map_put;
 mod map_remove;
 mod map_size;
@@ -38,6 +39,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(map::create(&package))?;
     commands.set(map_get::create(&package))?;
     commands.set(map_is_empty::create(&package)?)?;
+    commands.set(map_load_properties::create(&package))?;
     commands.set(map_put::create(&package))?;
     commands.set(map_remove::create(&package))?;
     commands.set(map_size::create(&package))?;